+++ tests/testthat/test-countwords.R
@@ -1,4 +1,5 @@
 test_that("You can count on count_words()", {
+  testthat::skip_on_os("linux")
   rmarkdown::draft(file = "test", template = "html_vignette",
                    package = "rmarkdown", create_dir = TRUE, edit = FALSE)
   testthat::expect_equal(iheiddown::count_words("test/test.Rmd"), 241)